# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Website Builder Platform API — bilingual (AR/EN)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/subscription/cancel:
    post:
      operationId: cancelSubscription
      tags: [billing]
      summary: Cancel subscription
      description: Cancels the user's active subscription and reverts to the free plan
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cancelledAt:
                    type: string
                    format: date-time
                required:
                  - success
        "404":
          description: No active subscription found
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create checkout session
      description: Creates a Stripe checkout session for subscribing or upgrading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /billing/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      description: Returns the user's invoice history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
...
  /billing/topup:
    post:
      operationId: topupCredits
      tags: [billing]
      summary: Top up credits
      description: Create a payment session to add credits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupRequest"
      responses:
        "200":
          description: Checkout URL for topup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - page
        - limit
        - total
...
    CreateCheckoutRequest:
      type: object
      properties:
        planId:
          type: string
          format: uuid
      required:
        - planId
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
          format: uri
      required:
        - checkoutUrl
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amountUsd:
          type: number
        status:
          type: string
          enum: [paid, pending, failed]
        description:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - amountUsd
        - status
        - createdAt
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    TopupRequest:
      type: object
      properties:
        amountUsd:
          type: number
      required:
        - amountUsd
...